validate box number in yup schema
